"Ok I found solution for Borderlands 3"

+++ LinuxGameStuffs/Borderlands3_Steam_patch.sh
@@ -1,5 +1,8 @@
 #!/bin/sh -e
 
+# **** Deprecated! use GloriousEggroll's 6.8-2 version!!
+
+
 # Setting up some variables to make the job easier...
 WORK_DIR="$HOME/.bd3_patch"
 ORIG_DIR="$CWD"
@@ -7,13 +10,16 @@ ORIG_DIR="$CWD"
 GAME_REF_NUM='397540' # Reference number for Borderlands 3 as suggested by Hovercraft
 
 # Mine is Pop! OS, which bases on Ubuntu. It can be different on other machines. Adjust this directory for your own situation
-PROTON_DIR="/home/$USER/.steam/debian-installation/steamapps/common/Proton - Experimental/" 
+#PROTON_DIR="/home/$USER/.steam/steam/steamapps/common/Proton - Experimental/" 
+#PROTON_DIR="/home/$USER/GameAndMedia/SteamLibrary/steamapps/common/Proton 5.13/" 
+
 
 # As I said, I usually gather all the Steam games together in ~/GameAndMedia directory. It will be different in your case. Change it!
 WINE_PREFIX_FOR_INSTALL="/home/$USER/GameAndMedia/SteamLibrary/steamapps/compatdata/$GAME_REF_NUM/pfx"
 
 # Game Data directory!
-GAME_BIN_DIR="/home/$USER/GameAndMedia/SteamLibrary/steamapps/common/Borderlands 3/OakGame/Binaries/Win64/"
+# GAME_BIN_DIR="/home/$USER/GameAndMedia/SteamLibrary/steamapps/common/Borderlands 3/OakGame/Binaries/Win64/"
+GAME_BIN_DIR="/home/$USER/GameAndMedia/SteamLibrary/steamapps/common/Borderlands 3/"
 
 # Good old die function
 die () {
@@ -27,7 +33,8 @@ die () {
 # apt install -y git cabextract 
 # if the script complains missing tools.
 [ ! -x "$(command -v git)" ] && die "We need git! ponk!"
-[ ! -x "$(command -v cabextract)" ] && "We need cabextract!"
+[ ! -x "$(command -v cabextract)" ] && die "We need cabextract!"
+( [ ! -x "$(command -v python2)" ] & [ ! -x "$(command -v python3)" ] ) && die "We also need Python2 or Python3!"
 
 # Dealing with work directory
 [ -d "$WORK_DIR" ] && rm -rf "$WORK_DIR"
@@ -43,7 +50,7 @@ if [ -d "$WORK_DIR" ]; then
   cd "$WORK_DIR" && \
   git clone https://github.com/z0z0z/mf-installcab && \
   cd "$WORK_DIR/mf-installcab" && \
-  PROTON="$PROTON_DIR" WINEPREFIX="$WINE_PREFIX_FOR_INSTALL" ./install-mf-64.sh
+  PROTON="$PROTON_DIR" WINEPREFIX="$WINE_PREFIX_FOR_INSTALL" taskset -c 0-3 ./install-mf-64.sh
 
   printf '\nmfplat.dll? will do!'
 
@@ -52,7 +59,7 @@ if [ -d "$WORK_DIR" ]; then
 fi
 
 # Cleaning up!
-printf 'Cleaning up everything!\n'
+printf '\nCleaning up everything!\n'
 cd "$ORIG_DIR" && rm -rf "$WORK_DIR"
 
-printf 'Jobs finished! Have fun with LOOTING!\n\n\n'
+printf '\nJobs finished! Have fun with LOOTING!\n\n\n'
